Add support for narrow integer types (#252)
Adds support for arrays of narrow integer types that protobuf doesn't support, specifically: int8, uInt8, int16, uInt16. Per the spec, int8[] is represented as the protobuf type bytes, and all the others are repeated int32. If a user passes in an int32 that is out of range for the real type, the wrapper code will throw an exception.