Add support for input varargs functions (#290)
Adds support for functions that take varargs as input parameters. (output parameters will be coming in another PR, but I want to get this in first to make sure it's on the right track) To test the behavior where there are too many parameters in a function, I added one real DAQmx function.