Add support for output varargs functions (#299)
Adds support for functions that use varargs to pass output. It builds on #290. To verify that it works with functions with many parameters this also adds a DAQ function `GetDigitalPullUpPullDownStates`.

To do this we now look more at the repeating parameters (the ones marked with `repeating_argument`), which now must match the fields in the repeated vararg.