Fix race condition in IVI dance (fixes #247) (#309)
Fixes a race condition if the size of a buffer changes between the initial sizing call and the "real" call afterwards.

This code assumes that if you pass in too small a buffer to a vanilla `ivi-dance` method it either returns the necessary size, or returns `kErrorReadBufferTooSmall`. For `ivi-dance-with-a-twist` methods, passing in too small a buffer will return `kErrorReadBufferTooSmall`.