update custom ivi-dance implementations (#312)
Per [this comment](https://github.com/ni/grpc-device/pull/309#pullrequestreview-734122133) on #309, update the custom implementations of `ivi-dance` to retry if the buffer is too small. (to avoid a race condition)

I looked at all of the custom service implementations and these were the only places that are using the `ivi-dance` or `ivi-dance-with-a-twist` patterns.