Determine array sizes for passed-in arrays in DAQ (#320)
Determines array size and passes that in automatically.

Note that this is just a proof of concept - I have not looked through the rest of the DAQmx functions to see which other ones could take advantage of this. This will hopefully let us get a sense for how hard this is and how much value it provides.