Coerce to narrow types inside of custom structs (#377)
Adds a checked conversion when converting to a narrow type that gRPC doesn't support. We previously did something similar with arrays of narrow types; this change extends the checks to happen inside of structs.

Note that this also adds checks to a few entry points in FGEN. This is technically a breaking change, as if a user was passing in values that were outside the range of an int16 the gRPC layer would previously do the call, while now it will return an error. However, previously the values that would be passed in to the C API would be whatever the compiler decided to do, so I think it is appropriate to call this a bug and fix it here.