Add custom types for RFSA service (#382)
Add all of the custom types for the RFSA service and the methods that use them.

Changes to codegen to support this:

Problem: SmtSpectrumInfo does not declare the standard _struct alias, which we use to disambiguate from the in-namespace proto type of the same name.
Solution: Extend additional_headers to declare required_by files so that it can be used in the library.
Problem: Several structs have reserved fields. If those are ever claimed, then use of the reserved name will be breaking.
Solution: Add support for include_in_proto to custom types and leave those fields alone.
Problem: SmtSpectrumInfo has a field called FFTSize. camel_to_snake turns this into f_ft_size.
Solution: Update camel_to_snake to detect pascal strings and pre-convert them.