Add missing DAQmx AnalogStatesWithOutputType functions (#408)
Adds missing functions.

GetAnalogPowerUpStatesWithOutputType() uses an output array sizing mechanism we haven't seen before - the size of the array is passed in by pointer, and the underlying API method sets that pointer to the actual size. If the actual size is bigger than what is passed in, the method returns an error. So this change adds a new mechanism passed-in-by-ptr.

Breaking change: Input arrays on Watchdog ExpirStates methods are given enum types, replacing int32 fields. Requires source changes, but does not break binary compatibility other than restricting valid values.