Fail workflow if files are dirtied during build (#423)
Update build_cmake.yml and build_nilrt.yml to fail if files are changed during build/codegen.

This ensures that our generated source code is up-to-date. This is important to (a) ensure that our review/tracking process for generated code is sound and (b) ensure that other workflow runs that operate on the generated source get the same inputs.