Add SockAddr type and converters to xnetsocket service (#575)
Add support for the nxsockaddr union type as a oneof type in grpc-device.

Our existing codegen does not support this type structure. To avoid adding a lot of complexity to codegen, I've bolstered 
the supports_standard_copy_convert concept and implemented a SockAddrHolder struct that does the heavy lifting.

Fixes AB#1871124.