Remove id from Session message (#784)
* Remove id from Session message type.
Motivation:
* Having both name and id is somewhat confusing
* It makes using the session in interactive tools like bloom RPC difficult, since oneof is a little complex to use.

I don't think ID should ever be used.  There are a lot of advantages to keeping the data on the wire "time invariant" or independent on server state.  For example we are trying to do batching to improve throughput when doing "cloud based measurements" as part of various LTIs and hopefully products.

Even if we remove id we can still maintain the current behavior by generating a name if the client does not specify.
Clients who use name should not need to make any changes to their code.
The wire format does not break for clients who use name so we don't even need to require clients change.
The wire format does not really break for clients that use id since they can't set it and our server will stop setting it.
Clients who use id should be treating Session as an opaque structure since there is nothing they can do with name.