Add nidmm_restricted service (#1188)
What does this Pull Request accomplish?
Adds a nidmm_restricted service to expose internal/private methods as needed. Specifically, this is pulling in the private methods that are needed to enable remote support for NI-DMM SFP in InstrumentStudio. See https://dev.azure.com/ni/DevCentral/_workitems/edit/3177197, https://dev.azure.com/ni/DevCentral/_git/ni-central/pullrequest/1055473.

Why should this Pull Request be merged?
This PR introduces the nidmm_restricted service to expose internal/private NI-DMM methods that are essential for enabling remote support in InstrumentStudio's Soft Front Panel (SFP). By integrating these private methods into the gRPC Device Server, we ensure that the necessary functionality is accessible by the Driver layer in InstrumentStudio. The following are the 5 APIs that are required for the remote support of NI-DMM SFP:

AttributeWasSetByUser
CachedReadStatus
FetchMultipointWithCaching
GetSessionsInformation
RequestPrivilege

What testing has been done?
Manual testing. The build was successfully completed on Windows. We are relying on the pipeline to validate the build on Linux.

Our validation was limited to confirming that the APIs are exposed to clients and are returning responses as expected. Since the gRPC Device Server acts as an intermediary layer that ultimately routes calls to the low-level DLL