Remove NI-DMM Private Header Files (#1219)
This pull request removes the nidmmPrivate.h and nidmmRemoteDebug.h that were used in nidmm. The functionality is still achieved through public header files.

Why should this Pull Request be merged?
This change is to ensure that we don't expose proprietary private entry points on NI gRPC Device Server.

What testing has been done?
Manual testing. The build was successfully completed on Windows. We are relying on the pipeline to validate the build on Linux.