Fix GOSI server-side implementation (#1222)
What does this Pull Request accomplish?
This pull request updates the NI-DMM GetOpenSessionsInformation API metadata and nidmm_restricted_service.custom.cpp to fix the Null Pointer Exception for NI-407x series devices.

Why should this Pull Request be merged?
This is a temporary workaround required for an internal release, and we will revert this change once the Exception is actually fixed in the Driver layer.

What testing has been done?
Relying on pipeline.